+++ 2023/kubernetes/images/api/.dockerignore
