+++ 2023/kubernetes/images/api/api/models.py
