+++ 2023/kubernetes/src/ptah/core/caching.py
