+++ 2023/kubernetes/src/ptah/core/image.py
