+++ 2023/kubernetes/src/ptah/core/shell.py
