+++ 2023/kubernetes/src/tests/test_image.py
